Fixed scenario in price fetching in which there the toFetch parameter is longer than 300 characters

+++ app/src/main/java/com/walletkeep/walletkeep/repository/AssetRepository.java
@@ -72,9 +72,9 @@ public class AssetRepository {
         CryptoCompareService.PricesResponseListener listener = new CryptoCompareService.PricesResponseListener() {
 
             @Override
-            public void onPricesUpdated(ArrayList<CurrencyPrice> prices) {
+            public void onPricesUpdated(ArrayList<CurrencyPrice> prices, Boolean delete) {
                 executors.diskIO().execute(() -> {
-                    database.currencyPriceDao().deleteAll();
+                    if (delete) database.currencyPriceDao().deleteAll();
                     database.currencyPriceDao().insertAll(prices);
                 });
             }
